Build for both EAP and Canary
