+++ .github/workflows/presubmit.yaml
@@ -18,7 +18,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        version: [4.1, 4.2, AF.3.1, 2020.3]
+        version: [4.1, 4.2, AF.3.1, 2020.3, 2021.1]
     steps:
       - name: checkout
         uses: actions/checkout@v2
