+++ build.gradle
@@ -16,12 +16,15 @@ buildscript {
     maven {
       url 'https://jetbrains.bintray.com/intellij-third-party-dependencies'
     }
+    maven {
+      url 'https://www.jetbrains.com/intellij-repository/releases'
+    }
   }
 }
 
 plugins {
-  id "org.jetbrains.intellij" version "0.6.5"
-  id "org.jetbrains.kotlin.jvm" version "1.3.70"
+  id "org.jetbrains.intellij" version "0.7.2"
+  id "org.jetbrains.kotlin.jvm" version "1.4.31"
 }
 
 repositories {
