+++ flutter-studio/src/io/flutter/utils/AndroidLocationProvider.java
@@ -5,7 +5,7 @@
  */
 package io.flutter.utils;
 
-import com.android.tools.idea.gradle.dsl.parser.BuildModelContext;
+import com.android.tools.idea.gradle.dsl.model.BuildModelContext;
 import com.android.tools.idea.gradle.project.model.GradleModuleModel;
 import com.android.tools.idea.gradle.util.GradleUtil;
 import com.android.tools.idea.projectsystem.AndroidProjectRootUtil;
