+++ product-matrix.json
@@ -10,10 +10,7 @@
       "dartPluginVersion": "201.7223.99",
       "sinceBuild": "201.7223",
       "untilBuild": "201.*",
-      "filesToSkip": [
-        "src/io/flutter/utils/AndroidLocationProvider.java",
-        "flutter-studio/src/io/flutter/actions/FlutterNewProjectAction.java"
-      ]
+      "filesToSkip": ["flutter-studio/src/io/flutter/actions/FlutterNewProjectAction.java"]
     },
     {
       "channel": "stable",
@@ -24,8 +21,7 @@
       "ideaVersion": "202.6987402",
       "dartPluginVersion": "202.8070",
       "sinceBuild": "202.7660.26",
-      "untilBuild": "202.*",
-      "filesToSkip": ["src/io/flutter/utils/AndroidLocationProvider.java"]
+      "untilBuild": "202.*"
     },
     {
       "channel": "stable",
@@ -36,8 +32,7 @@
       "ideaVersion": "2020.3.1.5",
       "dartPluginVersion": "203.6912",
       "sinceBuild": "203.6682.168",
-      "untilBuild": "203.7148.57",
-      "filesToSkip": ["src/io/flutter/utils/AndroidLocationProvider.java"]
+      "untilBuild": "203.7148.57"
     },
     {
       "channel": "stable",
@@ -50,6 +45,18 @@
       "dartPluginVersion": "203.6912",
       "sinceBuild": "203.7148.57",
       "untilBuild": "203.*"
+    },
+    {
+      "channel": "stable",
+      "comments": "IntelliJ 2021.1 EAP",
+      "name": "IntelliJ 2021.1",
+      "isUnitTestTarget": "true",
+      "version": "2021.1",
+      "ideaProduct": "ideaIC",
+      "ideaVersion": "211.6222.4",
+      "dartPluginVersion": "211.6222.2",
+      "sinceBuild": "211.6222.4",
+      "untilBuild": "211.*"
     }
   ]
 }
