--- src/io/flutter/utils/AndroidLocationProvider.java
@@ -1,2 +1,0 @@
-package io.flutter.utils;
-public class AndroidLocationProvider {}
